feat(app): create useQuestionsLevel hook
